+++ apps/jibsayo/src/app/api/shared/services/transactionService.ts
@@ -26,3 +26,15 @@ export const createTransactionId = (params: {
 
   return `${obfuscateKorean(safeRegionCode)}__${obfuscateKorean(safeAddress)}__${obfuscateKorean(safeApartName)}__${safeTradeDate}__${safeSize}__${safeFloor}__${safeTradeAmount}`;
 };
+
+export const createApartId = (params: {
+  regionCode?: string;
+  address?: string;
+  apartName?: string;
+}): string => {
+  const safeRegionCode = params.regionCode || '';
+  const safeAddress = params.address || '';
+  const safeApartName = params.apartName || '';
+
+  return `${obfuscateKorean(safeRegionCode)}__${obfuscateKorean(safeAddress)}__${obfuscateKorean(safeApartName)}`;
+};
