+++ apps/jibsayo/src/app/api/transactions/models/types.ts
@@ -1,4 +1,5 @@
 export interface TransactionItem {
+  transactionId: string;
   apartId: string;
   apartName: string;
   buildedYear: number | null;
