+++ apps/jibsayo/src/entities/transaction/models/types.ts
@@ -1,4 +1,4 @@
-import { TransactionsResponse } from '@/app/api/transactions/types';
+import { TransactionsResponse } from '@/app/api/transactions/models/types';
 
 export type TransactionItem = TransactionsResponse['list'][number];
